+++ NEWS.md
@@ -1,3 +1,10 @@
+# sagemaker.core 0.1.2.9001:
+## New Feature:
+* Initial implementation of jumpstart function and classes
+* Add `listenv` dependency to allow for mutable lists
+
+# sagemaker.core 0.1.2.9000:
+
 # sagemaker.core 0.1.1:
 ## Bug Fix:
 - Allow dynamic s3 uri to be parse by `s3_parse_url` https://github.com/aws/sagemaker-python-sdk/blob/a765512536fd5bd95733ce9aa497783a06cc8bdf/tests/unit/test_airflow.py#L256-L281